backend/pyproject.toml
[project]
name = "searchable-pdf-service"
version = "0.1.0"
description = "Convert scanned PDFs to searchable PDFs using Docstrange API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "pymupdf>=1.24.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
searchable-pdf = "app.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning:builtins"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
